Use BeaconValidator interface for beacon validations
