+++ ethereumj-core/src/main/java/org/ethereum/sharding/processing/validation/AttestationsValidator.java
@@ -17,7 +17,6 @@
  */
 package org.ethereum.sharding.processing.validation;
 
-import org.ethereum.sharding.crypto.DummySign;
 import org.ethereum.sharding.crypto.Sign;
 import org.ethereum.sharding.domain.Beacon;
 import org.ethereum.sharding.processing.db.BeaconStore;
@@ -35,7 +34,6 @@ import org.slf4j.LoggerFactory;
 
 import java.math.BigInteger;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 
 import static org.ethereum.sharding.processing.consensus.BeaconConstants.CYCLE_LENGTH;
@@ -46,7 +44,7 @@ import static org.ethereum.sharding.util.BeaconUtils.scanCommittees;
 /**
  * Validates block attestations
  */
-public class AttestationsValidator {
+public class AttestationsValidator implements BeaconValidator {
 
     private static final Logger logger = LoggerFactory.getLogger("beacon");
 
@@ -67,10 +65,6 @@ public class AttestationsValidator {
         this.repository = repository;
         this.sign = sign;
         this.rules = rules;
-    }
-
-    public static AttestationsValidator createDummy() {
-        return new AttestationsValidator(null, null, new DummySign(), Collections.emptyList());
     }
 
     /**
